Vector2O,Vector3O,Vector4OをVectorO.hに集約 それに伴うVector4Oを使用していたコードのエラー解消 MeshFilterの頂点情報を保存できるようにしてる おそらくFBXでもOBJでも受け取った頂点情報は全部CSV出力するので(未検証)調整が必要？

+++ EngineSystem/Material.cpp
@@ -79,7 +79,7 @@ void Material::DrawGUI() {
 
 void Material::ImportFile(std::vector<std::string>& tokens)
 {
-	m_Material.SpecularPower = std::stoi(tokens[2]);
+	m_Material.SpecularPower = (FLOAT)std::stoi(tokens[2]);
 	m_VertexShader = tokens[3];
 	m_PixelShader = tokens[4];
 	m_Texture = std::stoi(tokens[5]);
